requirements.txt
asgiref==3.2.7
Django==3.0.7
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
psycopg2==2.7.7
pytz==2020.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
